Move CPeriodState traits
